Import Quickfix based on Refactoring Infrastructure
Extracted some reusable bits from the refactoring action so we can reuse that in the quickfix.

The quickfix to add missing imports is now based on the refactoring library. At the moment, this doesn't have many benefits, but it gives us some options for future enhancements. For example, we could add the import in a enclosing scope.