Fixed #1000475. Completion is now hopefully working reliably in almost all scenarios. Further, completion finally also works when using infix annotation! (e.g., List(1,2) for<ask completion> now works!)
The fix consisted in avoiding calling 'askCompletion' inside the 
generic 'askOption' (that accepts a closure). 

The technical reasons seems that 'ask' sets 'interruptsEnabled' to
false, and 'signalDone' will not even look at the result tree,
therefore the targetPos will never match the resulting tree. It looks
like anything that relies on getting a typedTreeAt inside an 'ask'
will fail!