Fixed #1000524, #1000568 and #1000586.
ScalaStructureBuilder now correctly expose methods with generic signature to JDT. We use the Scala compiler for generating the java signature from a symbol, since that functionality is already there (call to erasure.javaSig(Symbol)).
JavaSig class encapsulate the call to erasure.javaSig.

In order to correct mapping of Scala Arrays (ticket #1000586) into Java ones, a new mapType(Type) method has been added to ScalaJavaMapper class. This method is also used to map types in absence of a java signature (which can occur if the method does not have any generic type in its declaration).

All snippet of code that were provided with the tickets have been added to the regression suite. I also had to make some correction to existing tests, particularly for the structurebuilder test named _traits_, as the test's oracle was wrong.

While with this commit we improve a lot interoperability of Java with Scala members, works still has to be done for mapping type parameters of classes. We have an open ticket for that (#1000625), but fixing it should be relatively easy now that we have a good infrastructure for retrieving generic java signature.