Don't rely on NameEnvironment for retrieving source folders and their output directories.
Additionally, cleaned up some duplicated code between 'outputFolders' and 'sourceOutputFolders'.
This changeset introduces a slight behavioral change, in that output directories are not created
if they don't exist. This does not seem to pose any problems. Creating such folders locks the
workspace, and if this happens during a build, it may lead to deadlocks.

Fixed #1000662.