General hardening of Scala compilation units.
Indexing, override indicators and the structure builder
are going to report a crash in the compiler only once, then silently ignore it until the file is
re-open. Certain exceptions, such as StackOverflowError, are reported by Eclipse in a dialog, and that
can make editing impossible if the presentation compiler keeps crashing (because of a bug, of course).

Disabled indexing and override indicators for Scala files in projects without the Scala nature (because
that means there's no Scala library on the class path).

Fixed #1000535 *again*.