Created test for ticket Re #1000692, which is currently disabled because the issue needs to be fixed in the presentation compiler (I tried to reproduce the problem with a presentation compiler test, but so far I did not succeeded in isolating the issue outside the IDE. Though, I'm sure that the issue described in this ticket has to be fixed in the presentation compiler.).
While creating the test I had to refactor a bit class 'HyperlinkDetectorTester'. Specifically, I needed to extract the functionality used to test hyperlinking in a scala file, as the issue described in this ticket becomes apparent only *after* a hyperlink request is issued on a specific class' symbol.

Last, I also changed CustomAssertion to have a self type 'TestProjectSetup', as in this way the current 'project' does no longer need to be passed to 'assertNoErrors'.