Updated ScalaStructureBuilder, as it is responsible of correctly exposing module's inner classes to JDT. The issue was that module's inner classes were exposed as member of the $ module class, which is wrong, since the scala compiler expose them as inner classes of the module's companion class.
I moved the code that was responsible of traversing the Scala AST Tree into a new class 'TreeTraverser', that takes a new field 'module2innerClassDefs' used to keep track of inner classes that are defined in a module. This solution has the advantage of avoiding possible memory leaks that could result from not correctly cleaning the 'module2innerClassDefs' map.

Fixed ticket #1000678.