The ScalaMethodVerifierProvider needs to access the active project to query the presentation compiler and retrieve the symbol associated to a MethodBinding.
To retrieve the active project we need to use the Eclipse Workbench, which is a UI component. Consequently, I had to
ignore all tests related to the ScalaMethodVerifierProvider, as they would fail in headless mode.

The strategy we used in the past to retrieve the project was simply broken. In fact, it was crashing when the MethodBinding
was coming from a binary classfile (check Re #1000660 for a practical example).

I'll work on re-enabling them but I'd like to push this forward as it Fixes #1000660, #1000715.

Iulian is right, if the MethodBinding is coming from a binary file, then we don't need to run it through the
ScalaMethodVerifierProvider, because all the mixed-in methods have been added (therefore Java compilers sees
the correct types).