Ticket Re #1000741 raised the question of whether we needed an open editor to check if the MethodBinding passed to the ScalaMethodVerifierProvider is indeed abstract (the reported NPE exception was not generating any issue, it was simply creating lot of noise in the Eclipse Error Log view).
The answer is no, and I updated the code accordingly. Now, we check if the MethodBinding belongs to a trait only
if the MethodBinding is declared in a Scala file (we look at the file's extension to infer this). That makes much
more sense and allowed to re-enable all tests in jcompiler package.

Any exception occuring in this code is still being logger in the Error Log view, that because I believe that now
the code should work as expected (and if it doesn't, I'd prefer not to be silent so that the issue is reported and
we can look into it).