One step closer. Eclipse compiler to follow sbt requests.
Added code to update on the fly the delta used by the Eclipse Java compiler, so it compiles
the file requested by sbt.

It works fine for RE #1000607 on my system.
For RE #1000388, it is only part of the solution. It mostly works but sometimes sbt doesn't request the recompilation of the
dependent Java class.
(cherry picked from commit fbfffa2883238fcb2d3ff31d72b2c1ff75cbdc71)