Disabled test because it is making fail the build for 2.8 compiler.
As a matter of fact, it is really only by chance that with 2.9 the test is passing, and
a bigger issue is lying behind this test. Look at ticket Re #1000761. The problem is that
we are not correctly exposing to JDT private values defined in a trait , i.e., the exposed
getter does not match the name of the method in the binary, which is clearly an issue (for
instance, I'm pretty convinced that the spurious error _The hierarchy of the type XXX is
inconsistent_ that can sometimes be shown in the Java editor, is likely to be linked to the
very same problem).

Since we plan a major re-work of the structure builder for 2.1 I believe it is better to tackle
this at that point. no review.