Removed the AJDT weaving dependency so that we are more compatible with Spring IDE and AJDT.
Copied the corresponding code from the AJDT plugin
to our aspects code. Kept the same two extension points.

Changed the update site to show only the Scala IDE
feature (binary and source). The weaving feature is
a nested feature in the SDT feature, because it needs its
own license.

Added LICENSE files. The Scala license for the project,
and the EPL for the org.scala-ide.sdt.aspects bundle.

Fixed #1000780.