Added 'org.eclipse.jdt.groovy.core' in Eclipse-SupplementBundle section of MANIFEST. Fix #1000798
To correct #3135 we removed the AJDT dependency from the Scala IDE. An
undesired side-effect of this change was that the Scala IDE was no longer
usable if the Groovy IDE was also installed.
The problem was that our (aspect) class responsible of intercepting the call to
create a ScalaSourceFile was no longer used, that because our class was
targeting the 'org.eclipse.jdt.core' bundle, which is not loaded when the
Groovy IDE is installed.
The fix was to edit the MANIFEST's Eclipse-SupplementBundle section of our
'org.scala-ide.sdt.aspects' project, and make sure that our aspect class is
called also when the call is performed from the 'org.eclipse.jdt.groovy.core'
bundle (which is the Groovy IDE core bundle). The AJDT plugin does exactly
that, and that is why this problem was a side-effect of removing the AJDT
dependency from the Scala IDE.

(no review)

cheery-picked from 3b477dd3c3a9ecab9e47343534be0b471a9cbcd4