Perform code-completion and import-adding in one step.
We need to to code-completion and import-adding in one step, otherwise
we risk that adding the import fails because the code-completion has
changed the underlying file and the AST positions are out of date for
the refactoring-library.

Fixes #1000854.

This is a backport of dcec1f3df5f572c036f7ec9e6dd789e2ffeaa44c.