Integrated Iulian's review from pull #64.
As suggested, I moved the Logging settings from the Scala Diagnostic dialog to
the Eclipse Scala IDE Preferences page.

Also, I removed the "plugininfo" option, which it was used to enable debug
information in the ScalaIndexBuilder. As a matter of fact, the amount of
logged information can now be easily controlled through Log Levels.