Refined places where askOption need to be used in ``SymbolClassifier``.
+ Made sure that symbols are managed only within the Presentation Compiler Thread.
+ Prevented NullPointerException during SemanticHighlightingReconciliation.createSemanticDecorationManagers (as the editor may be null).
+ Followed Mirko's suggestion (https://github.com/scala-ide/scala-ide/commit/11fd6167587d24bd52fe3952220b87278677691f#commitcomment-972940), i.e., removed `TreeCreationMethods` mixin from `SymbolClass ification` class.

Also, in SymbolClassification, there is no need to store the source file
content in a string.  This is really needed only if the user enabled
`useSyntacticHints`. Not doing it by default can save a few cycles,
particularly for big files.

Further, I've added SymbolClassifierTestSuite to the TestSuite, so that the
tests are run at every build.