For no apparent reason, using `ScalaPlugin.defaultScalaSettings` is not the same as `new Settings` in `SettingsCleanup` class, and I honestly really can't see a reason (if anyone has an idea, please don't be shy).
The issue with using `ScalaPlugin.defaultScalaSettings` is that SBT may fail
compiling sources that depends on the continuations plugin. For instance,
assume you have a `plugins` folder (located in ${PATH_TO_PLUGINS_DIR})
containing the continuations.jar plugin, and in the project's settings the
following options are passed:

  -P:continuations:enable
  -Xplugin ${PATH_TO_PLUGINS_DIR}/continuations.jar
  -XpluginsDir ${PATH_TO_PLUGINS_DIR}

The SBT compiler will fail to compile the sources because it can't locate the
continuations plugin.  While, when using `new Settings`, it all works fine.

I've add a FIXME note to remind us of the problem.