Made sure that logger is initialized before being used.
When starting the IDE, you may have noticed the following message in the
console:

  log4j:WARN No appenders could be found for logger (ScalaPlugin).
  log4j:WARN Please initialize the log4j system properly.
  log4j:WARN See http://logging.apache.org/log4j/1.2/faq.html#noconfig for more info.

Which was caused by some eager fields in ScalaPlugin. Since the logger is
initialized when calling `ScalaPlugin.start`, eager fields in `ScalaPlugin`
should not try to access the logger or the above warning message is produced.

To fix the issue, the eager fields have been transformed into lazy ones.