Temporarily remove the SysOut/Err redirects while attaching the Console Appender to the set of log appenders.
It has been noticed that enabling the console appender was generating a stackoverflow exception. The reason
was that the standard output/error is being redirected to the logger, and attaching a console appender was
causing a recursion.

To avoid the recursion it is sufficient to disable the standard output/error redirects while attaching the
console appender.

Fix #1000903.

Integrated comments from pull #71.