Passing to -Xplugin a path containing whitespaces is finally parsed correctly.
Up until now, passing to -Xplugin a path containing whitespaces resulted in
*no* plugin being loaded. That because
``MultiStringSetting#tryToSetFromPropertyValue`` was splitting the passed value
at whitespace occurrences, which was of course resulting in a mess.

The pull request https://github.com/scala/scala/pull/235 in the Scala compiler
has fixed this issue (patch available in both 2.10 and 2.9.2). In short, now
commas (``,``) are used to split -Xplugin value (when more than one plugin is
passed to it). Some adaptation on our side was also needed for correctly
handling commas as the splitting character.

This commit also fix a issue with the way we used to hande -Xpluginsdir value.
Read the big comment in ScalaPlugin.defaultScalaSettings for more information
about this.

As a side note, enabling continuations should now work as described in
http://scala-ide.org/docs/tutorials/continuations-plugin/index.html.

Fixes #1000917.

Integrated @skyluc code review (pull #75) and added a couple of new tests to
check that compilation of a source requiring the continuations plugin enabled,
works as expected.
(cherry picked from commit cbfd82d874f998b7726fec895ad4d83194ad1e5b)

Conflicts:

	org.scala-ide.sdt.core.tests/src/scala/tools/eclipse/testsetup/TestProjectSetup.scala