Open Declaration via the context menu now works as expected.
When the user right clicks on a element and select "Open Declaration" in the
context menu, an instance of ``OpenAction`` is used to resolve the binding and
jump to the declaration.

Because the Eclipse API does not expose an extension point for this action, I
had to create a custom extension point (using AJDT), which allows weaving in to
JDT and can be used to intercept the creation of ``OpenAction``, if it
originates from a Scala editor.

This seem to be the only solution, which does not involve creating our own
Scala Editor. Hopefully, we will be able to remove this custom extension point
once we implement the Scala Editor (this is planned for Milestone 2 of the
Scala IDE Helium).  Using AJDT for customizing the behavior of ``OpenAction``
was also suggested by Andrew Eisenberg in the following StackOverflow question:

	http://stackoverflow.com/questions/1882053/how-can-i-define-a-custom-action-for-open-declaration-shortcut-f3-in-eclipse

Fix #1000920.
(cherry picked from commit d43126db924b6f0d138588652db491e0de856b3a)