Correctly handle preferences change events
Some Java preferences get set the first time a code template is
used. This preference changes trigger UI updates, but the Java editor
(base of the Scala editor) does not execute these UI updates in the right
thread.
Forcing those preference change to be run in the UI thread.
Fixes #1000738