Semantic Highlighting symbols' classification should not skip synthetic symbols.
In the attempt of optimizing semantic highlighting, synthetic symbols were
being skipped. That seemed to be a valid optimization, since synthetic symbols
are associated to compiler injected definitions (which are hence not visible in
the source code). However, user code may be use the synthetic definitions, and
therefore it is not safe to filter away synthetic symbols during
classification.

For instance, in the following example

```
case class CaseClass(n: Int) {
    CaseClass(42) // <- this reference the companion object, which is a synthetic symbol
}
```

``CaseClass(42)`` is actually referencing the companion object of the class
``CaseClass``. The companion object has not been declared in the source, hence
it is injected by the compiler, and a synthetic symbol will be used to
reference it.

On the other hand, it looks safe to skip anonymous function's symbols.