Correctly retrieving the symbol associated to an ``import`` clause.
I believe you will agree that finding the *right* symbol it's a interesting
challenge in the case of an ``import`` clause. In a nutshell, the only way to
know whether the imported ``name`` is referencing a class is to check if the
symbol associated to ``expr.tpe.member(name.toTypeName)`` actually ``exists``.
If it doesn't then the ``name`` must be referencing a module.

There is one more issue with the way we resolve ``import`` clause: we are not
correctly manipulating clauses like ``import scala.Option._``. This still needs
to be fixed.