Corrected classification of ``var``s.
A ``var`` that gets accessed through its getter (remember, a ``var`` gets
always compiled into a getter and a setter by ``scalac`` - uniform access
principle) must be classified as a variable (either TemplateVar or LocalVar).

The difficulty is that when a ``var`` gets accessed through its getter, the
symbol handle we have in the ``Tree`` is a ``MethodSymbol`` for the getter.
Therefore, to decide if behind a getter sits a ``val`` or a ``var`` we need to
check if the symbol's owner (i.e., the getter's owner) contains a setter method
that matches the getters' name (this can be done using the compiler method
``nme.getterToSetter``).

Fixed #1000958.