Optimization: Reduce number of time a Scala source is marked as changed
Marking a Scala source as changed may trigger reconciliation, which can
be a costly operation at this time since as the semantic actions are run
in the reconciler thread.

More importantly, a source should be marked as changed only if some change
really occurred, that's why we now check if the source's ``delta`` has the
flag ``IJavaElementDelta.F_CONTENT`` set.