Removed one source of quadratic behavior in semantic highlighting.
Added debug timings, and replaced a call to `Map.mapValue` with `map`. The main reason is that `mapValue` does not cache any results, and in this case the resulting map is called thousands of times with the same key. See comment in code for the exact source of quadratic behavior. This shaves off a good couple of seconds when highlighting `Typers.scala` (>5000 loc) on my laptop.