ImplicitHighlight should not force re-typechecking of source
Forcing typechecking of a source is an expensive operation for files with more
than a few hundreds lines (e.g., take the Typer source in the Scala project,
which is about 4KLOC).

Moreover, forcing typechecking is often not necessary, as it is the job of the
background compiler to decide when and how a source should be typechecked.

It turns out that the ``ImplicitHighlightingPresenter`` was calling
``compiler.body(source)`` to find all implicit conversions. ``compiler.body``
forces re-typechecking, which is bad.

The fix was simply to call ``compiler.loadedType`` instead, which perform a
full typechecking of the passed source only if the source is not yet loaded.