Implemented hyperlink to implicit
When an implicit conversion is applied, click COMMAND and hover the highlighted
binding, a contextual menu will open giving you a chance to navigate to the
implicit declaration (the default is to navigate to the binding's declaration).

At the moment it is unfortunately not possible to create UI-less tests for this
feature, because hyperlink to implicits are stored in the editor's annotation
model (this is done by the ImplicitHighlightPresenter).

Fix #1001002