Fixed hyperlink implicit for (Tree) ``Literal``
On the one hand, a ``Literal`` (e.g., an integer value, a string value, ...)
never have a declaration hyperlink. On the other hand, a Literal may have a
implicit applied to it, and therefore an implicit hyperlink should be created
for it.

Hence, it makes sense to separate discovery of declarations and implicits
hyperlinks.