Fixed Hyperlinking in classfiles regression
CMD+hover on classfiles wasn't working anymore after the recent addition of the
``Hyperlink to Implicit`` functionality.

The reason was that ``Hyperklink to Implicit`` currently needs to go through
the editor's annotation model to know if an implicit conversion is getting
applied. Apparently, the ``IEditorInput`` of a ``ScalaClassFile`` (which is an
``InternalClassFileEditorInput``) does not have an ``IAnnotationModel`` attached
to it, and that caused the ``ImplicitHyperlinkDetector`` to throw an exception,
which was the reason why CMD+hover was no longer working properly.

The fix was simple enough. In ``EditorUtils.getAnnotationsAtOffset`` returns now
an empty ``Iterator`` if the editor's annotation model is ``null``.