Moved source generators to Source menu, properly handle refactoring preparation errors.
Generate hashCode and equals and Introduce ProductN trait now show up in the Source context menu. This is done programmatically in ScalaSourceFileEditor#editorContextMenuAboutToShow, so they are removed from plugin.xml.

When a refactoring fails in the preparation step this is now properly handled, the error message is presented in a dialog.