Clean up and simplification of the poms
Made the dependencies of library, compiler and sbt optional, so they don't have to be excluded later.
Moved the 'allowConflictingDependencies' option to the projects which require it (feature and site).
It is needed because of the mix of versions for the elements coming from the m2 repo.
Removed the dependency to aspectjrt (not needed)
Cleaned-up an invalid property definition in sbt.full.library.
Created properties for all plugin version numbers
Fixed typo
Fixed git location
Added parameter to the p2 repositories ids