Create fatal errors when a refactoring's initial check yields an error.
Until now, we created normal errors, which were shown in a wizard page,
and because they were non-fatal, the refactoring still tried to execute,
resulting in exceptions in the log file. Now we use fatal errors, which
immediately abort a refactoring and show the error message in a window.

Fixes scala-refactoring #76