Handle invalid target packages in the Move participant.
When the target package of a move refactoring cannot be found (e.g. because
the target is not a source folder), we abort with a warning. The user can
then still move the file, but he will be warned that the source won't be
changed. The JDT doesn't warn the user but simply moves the file without
refactorings its content.