More reliable refactoring in multi-file situations.
Stop using lazy vals for caching compiler objects, because this leads to
problems with the compiler's symbols. In detail: when starting a refactoring,
the current file is type checked and used for initial condition tests.
Then depending on the refactoring, more files are type-checked and added
to an index. We used lazy vals to hold the results of the initial
checks, e.g. the symbol of the selected method, but this then leads
to problems when this "old" symbol is used in the index that was built with
the newer symbols. In short, it gave me many headaches, and the tiny
performance improvements were not worth it.

Also fixes a bug with the project wide index which only searched for
types and not other elements.

Fixes #1001059.