Refactoring: Moved logic for opening a source in ``ScalaSourceFile``
* Added three new methods ``forceReload``, ``reload`` and ``discard`` in
``ScalaSourceFile``, and moved the logic from ``ScalaPlugin`` to
``ScalaSourceFile ``. This was done also for testing purposes (next
commit), as I needed a way to exercise the opening of a source file
(without having to duplicate the logic).

* As a side remark, I wonder if there isn't an already existing method in
  ``ScalaSourceFile`` that we should override to run the logic that is in
``reload/discard``. I though ``open(IProgressMonitor)`` could be the right one,
but that method doesn't seem to be called when opening a source file in the
editor.