Corrected ``MainClassVerifier`` (Re #1001096)
The logic in ``MainClassVerifier`` was partially broken. Indeed, the logic used
for checking that the main unit's package declaration matched its physical
location (rule 3. in the comment) was:

  1) plain wrong, and
  2) not neeeded because rule 2. was already taking care of it.

Therefore, all code related to rule 3. was removed, and the test provided with #1001096
has been added to the regression suite.

Re #1000541

Fix #1001096