Warn the user if JDT Weaving is OFF
At startup, check that JDT Weaving is enabled. If it's not, then
warn the user through a dialog box and programmatically
enable JDT Weaving (and restart the workbench) upon confirmation.

Note that there is no way for the user to silent the dialog. Each time the
Scala IDE plugin is activated the check will be performed, and if JDT
Weaving is OFF the above mentioned message will be displayed. If the
user doesn't want to see the dialog, it will have three options:

* Enable JDT Weaving
* Disable the Scala IDE core bundle (using the OSGi console)
* Uninstall the Scala IDE plugin

Fix #1001104
(cherry picked from commit 64aa951316b0f717172906f5a2542e78696fd4b2)