Use textual instead of Java search to find files to index.
Instead of using a Java search for files that need to be included
in the refactoring index, we use a simple textual search. The problem
with the Java search was that it didn't find all occurrences of a term,
e.g. when a file contained only an import of a type, it wasn't found
by the Java search and thus not indexed and consequently not part of
the rename refactoring.

Fixes #1001117