Truncate too long build error messages displayed in resources' marker
Markers are a general mechanism for associating notes and meta-data with
resources.  The Eclipse environments dictates that the text message passed to a
marker should not be bigger than 21000 chars, or an assertion failure saying

> Marker property value is too long

is thrown on the user's face.

The actual logic for resizing the build error problem passed to a marker
already existed in two places. While the message passed to the created marker
was correctly truncated in ``FileUtils.buildError``, it was not in
``ScalaProject.buildErrors``.

As part of this commit, I've centralized the creation of resource markers in
``MarkerFactory``, and build problem markers are now created using
``BuildProblemMarker`` (it's a subclass of ``MarkerFactory```).

More cleaning up could be done so that all markers are created only through
sublcasses of `MarkerFactory``. Though, such a big refactoring seem too risky
at the moment, as this commit will be merged in ``release/2.0.x`` branch.

Fix #1001107
(cherry picked from commit be013c2cb87c695b3cf1048b0280a64c73144ef6)

Conflicts:

	org.scala-ide.sdt.core/src/scala/tools/eclipse/buildmanager/BuildReporter.scala
	org.scala-ide.sdt.core/src/scala/tools/eclipse/buildmanager/refined/EclipseRefinedBuildManager.scala
	org.scala-ide.sdt.core/src/scala/tools/eclipse/buildmanager/sbtintegration/EclipseSbtBuildManager.scala