Fixed exception occurring when deleting a Scala project
Issue Description
.................

The issue was occurring only when a source was open in the editor and the
corresponding project was deleted. Apparently, when deleting a project, any of
its opened sources can still be reconciled. The problem with running the
reconciler on a virtual source (i.e., the source does no longer exist in the
filesystem) is that the reconciler participants (e.g., semantic and implicit
highlight presenters) expect the source file to phisically exist. When that's
not the case, a ``java.lang.IllegalArgumentException`` was thrown to the
user's face.

Implemented Solution
....................

The current fix is to check in
``SemanticHighlightReconcilation.beforeReconcile`` that the compilation unit is
linked to an existing project (i.e., the project actually exist in the
filesystem).

Improvement
...........

Furthermore, when a Scala project is about to be deleted, the Scala project is
now disposed (i.e., the presentation compiler is shut down). This was not
needed to fix the issue described in this commit, but it seemed a valid
optimization.

Tests
.....

Unfortunately, the changes cannot be tested in a UI-less environment, hence no
tests.

Fixed #1001058